Aggregate dataframe on adoption_date and covariates.

NB: Dataframe Y should presorted in non increasing order by adopt_date.

Usage

aggregate_Y(Y, covs, time_cols)

Arguments

Y

Dataframe. The dataframe should contain 'adopt_date', 'popwt', covariate and time columns with outcomes.

covs

Character vector. The covariate columns to be used in aggregation.

time_cols

Character vector. The time columns of Y dataframe with outcomes.

Value

     List. The list contains:

Y: Dataframe. The dataframe has 'Group', 'popwt' and time columns. 'Group' column is concatenation of 'adopt_date' and covariate columns.

N: Numeric. The number of rows of the Y dataframe after aggregation.