Estimates the DGP parameters used in the placebo studies of the Sequential Synthetic DiD paper.

Estimates the DGP parameters used in the placebo studies of the Sequential Synthetic DiD paper.

Usage

estimate_dgp(Y, assignment_vector, rank, pi_singular_vec, never_treat)

Arguments

Y

: Dataframe. It is a dataframe of outcomes.

assignment_vector

: Binary/boolen vector. It is a Nx1 vector of treatment assignments.

rank

: Integer. The rank of the estimated signal component L.

Value

A list with the following elements:

F: Numeric matrix. The FE of the

M: Numeric matrix. The interactive FE of the

Sigma: Numeric matrix. The noise covariance matrix.

sigma_2: Numeric. The noise variance: the number which is located on the main diagonal of the covariance matrix.

pi: Numeric vector.

ar_coef: Numeric vector. AR(2) model coefficients for the covariance matrix Sigma. and an element ar_coef with the AR(2) model coefficients underlying the covariance Sigma

rownames: Character vector. The row names of Y matrix.

colnames: Character vector. The column names Y matrix.