Estimate the DGP for CHC simulation.

The function estimates the DGP (interactive FE and noise) and adjusts the data to fit our needs:

Increase the number of units in each cohort

Adjust the ratio of noise variance to interactive FE variance

Usage

estimate_dgp_chc(panel, signal_share = 0, scale = 4, rank = 5)

Arguments

panel

List. The panel on the unit level.

signal_share

Numeric. The share of interactive FE

scale

Integer. The number of units in each cohort is rescaled by this factor to ensure our cohorts are large enough for the asymptotics to be relevant.

rank

Integer. The rank of interactive FE to extract from the data.

Value

        List.    The list contains:

L Dataframe. The dataframe contains interactive fixed effects.

X Dataframe. The dataframe has auxiliary data like adoption date, population weights and covariates.

sigma_2 Numeric. The noise variance on the unit level. The noise is supposed to be homoskedastic and does not have any autocorrelation.

N Integer. The total number of units after rescaling.

T Integer. The number of time periods.