Computes the cohort level sequential synthetic diff-in-diff or diff-in-diff estimate.

The outcome variable is aggregated on cohort or covariate level.

Usage

estimation_cohort(
  Y,
  W,
  coh_weights,
  penalty_func = default_penalty,
  s2 = NULL,
  type = "sdid",
  compute_var = FALSE,
  N0 = 1,
  fast_mat_inv = FALSE
)

Arguments

coh_weights

: Numeric vector. The Nx1 cohort weights vector (the number of units in cohorts or the cohort population).

s2

: Numeric. The upper bound estimate of noise variance.

type

: Character. Type of the estimator should be sdid or did.

N0

: Integer. The number of control (never-treated) units.

fast_mat_inv

: Bool. If TRUE, .

Y_avg

: Numeric matrix. The N x T aggregated matrix of outcomes.

W_avg

: Binary or boolean matrix. The N x T matrix of treatment indicators.

compute_var:

Bool. If TRUE, computes the asymptotic variance of the base estimator in homoscedasticity case.

Value

tau_hat: Numeric matrix. The N x T matrix of raw (not aggregated on lag level) treatment effects of treatment effects.