Aggregate panel to cohort level or prepare for the unit level.

The function prepares the data for sequential_estimator. For "unit" level the function removes adoption date column in Y and W. Y and W can be both matrices without adopt_date column. For "cohort" level the function aggregates the outcome and population variable by adoption date and covariates.

Usage

prepare_wide(panel, level = "cohort", boot = FALSE)

Arguments

panel

List. The list should contain:

Y Dataframe or numeric matrix. This is a wide panel dataframe of outcomes with adoption date column being the first one. Y can be a matrix when the estimation is conducted on the unit level.

X Dataframe. This is a dataframe with auxiliary data like adoption date, population weights and covariates. or

Y_wt Dataframe. This is Y dataframe with merged X dataframe. It contains

W Dataframe, numeric matrix or NULL. This is a wide panel dataframe of treatment indicators with adoption date column being the first one. NULL can be passed when boot is TRUE. Numeric matrix can be passed when the estimation is conducted on the unit level.

level

Character. The level should be unit or cohort. The data is not being aggregated when the level is "unit".

boot

Bool. If TRUE, W_avg matrix is not calculated to speed up the bootstraping for standard error estimation.

Value

panel_avg List. The list contains: -Y_avg: Numeric matrix. The N x T matrix of outcomes. -W_avg: Binary or boolen matrix. The N x T matrix of treatment indicators. The matrix has a stairlike structure with treated cells being in the bottom. -coh: Numeric vector. The N x 1 cohort weights vector (the number of units in cohorts or the cohort population).