Randomize treatment to n units with probability pi then if the number of treated units is zero, assign treatment to one unit uniformly at random and if the number of treated units exceeds a cap, remove treatment uniformly at random so it is exactly that cap

Randomize treatment to n units with probability pi then if the number of treated units is zero, assign treatment to one unit uniformly at random and if the number of treated units exceeds a cap, remove treatment uniformly at random so it is exactly that cap

Usage

randomize_treatment(pi, N, N1)

Arguments

pi

: Numeric vector. The randomization probabilities.

N

: Integer. The total number of units.

N1

: Integer. The cap on the number of treated units.

Value

assignment_sim Binary vector. The vector is of length N, with ones indicating assignment to treatment.