Simulates data from DGPs used in the placebo studies of the synthetic difference in differences paper.

Simulates data from DGPs used in the placebo studies of the synthetic difference in differences paper.

Usage

simulate_dgp(params, N1, a_min, permute_pi = FALSE, as_df = TRUE)

Arguments

N1

: Integer. A cap on the number of treated units.

a_min

: Integer. The number of treated periods.

permute_pi

: Bool. If TRUE, pi is permuted, so the treatment status is random.

as_df

: Bool. If TRUE, returns Y and W as dataframes instead of matrices.

parameters

: List. The list contains dgp parameters (F, M, Sigma, pi) as output by estimate_dgp function. F and M are numeric matrices.

Value

List with the following elements:

Y: Numeric matrix or dataframe.

W: Binary/boolen matrix or dataframe.

Sigma: Numeric matrix.

cohorts: Numeric vector. The number of control units N0, and the number of control periods T0. The first N0 rows of Y are for units assigned to control, the remaining rows are for units assigned to treatment.