Converts a data set into wide panels.

Converts a data set into wide panels.

Usage

to_wide(
  panel,
  unit = 1,
  time = 2,
  outcome = 3,
  treatment = 4,
  population = NULL,
  contr_covs = c(),
  treat_covs = c(),
  never_treat = "2500",
  sort = TRUE
)

Arguments

panel

Dataframe. The dataframe is utilized to

unit

Numeric or character. The column number or index that corresponds to the unit identifier.

time

Numeric or character. The column number or index that corresponds to the time identifier

outcome

Numeric or character. The column number or index that corresponds to the outcome identifier.

treatment

Numeric or character. The column number or index that corresponds to the treatment indicator.

population

Numeric, character or NULL. If not NULL, the column number or index that corresponds to the population identifier.

contr_covs

Numeric or character vector. If the length of the vector is not 0, the vector coresponds to the covariates for control (never-treated) units.

treat_covs

Numeric or character vector. If the length of the vector is not 0, the vector coresponds to the covariates for ever treated units.

never_treat

Character. The time index that indicates never-treated units.

Value

panel List. The list contains:

Y: Dataframe. This is a wide panel dataframe of outcomes with adoption date column being the first one.

W: Dataframe. This is a wide panel dataframe of treatment indicators with adoption date column being the first one.

X: Dataframe. This is a dataframe with auxiliary data like adoption date, population weights and covariates.