Computes the unit or cohort level sequential synthetic diff-in-diff or diff-in-diff estimate.

Basically, this function is a wrapper which calls either estimation_cohort or estimation_unit function. Then the function aggregates the treatment effect by length of the exposure.

Usage

sequential_estimator(
  panel_avg,
  level = NULL,
  s2 = NULL,
  type = "sdid",
  aggregate_effect = aggregate_inv_did_var,
  penalty_func = default_penalty,
  compute_var = FALSE,
  fast_mat_inv = FALSE
)

Arguments

panel_avg

: List. The list should contain:

Y_avg: Numeric matrix. The N x T matrix of outcomes.

W_avg: Binary or boolen matrix. The N x T matrix of treatment indicators. The matrix has a stair-like structure with treated cells being at the bottom.

coh_weights: Numeric vector. The Nx1 cohort weights vector (the number of units in cohorts or the cohort population).

X: Dataframe. The dataframe is not aggregated to cohort and covariate level. It contains auxiliary data like adoption date, population weights and covariates.

level

: Character or NULL. If NULL, the level is taken from panel_avg list. The level should be unit or cohort.

s2

: Numeric or NULL. The upper bound estimate of noise variance.

type

: Character. Type of the estimator should be sdid, did or both.

aggregate_effect

: Function. The function aggregates the treatment effect for every lag by weighting the effect for every cohort. The function should contain the following arguments:

tau: Numeric matrix or array.

W: Numeric matrix.

coh_weights: Numeric vector.

N0: Integer. The number of control (never-treated) units.

N: Integer. The total number of units.

compute_var

: Bool. If TRUE, computes the asymptotic variance of the base_estimator when the noise is homoscedastic and there is no autocorrelation. This parameter should be set to TRUE if aggregate_effect function utilizes the asymptotic variance (only aggregate_inv_var).

fast_mat_inv

: Bool. If TRUE, matrix equation (H * w = - g) is solved with solve_opts::fast flag in Armadillo.

penalty_func:

Function. You can create your own one and pass it as a parameter. The function should have the following arguments:

s2 : Numeric or NULL. The estimated noise variance.

N : Numeric.

Value

             sequential_estimator